重构，拆分 Preprocessor.php，移动到 src/sapi 目录
