+++ composer.json
@@ -5,7 +5,7 @@
   },
   "autoload": {
     "psr-4": {
-      "SwooleCli\\": "sapi"
+      "SwooleCli\\": "sapi/src"
     }
   },
   "require-dev": {
