+++ sapi/src/Preprocessor.php
@@ -7,216 +7,6 @@ use MJS\TopSort\ElementNotFoundException;
 use MJS\TopSort\Implementations\StringSort;
 use RuntimeException;
 
-abstract class Project
-{
-    public string $name;
-    public string $url;
-    public string $path = '';
-    public string $file = '';
-    public string $md5sum = '';
-
-    public string $manual = '';
-    public string $homePage = '';
-
-    public string $license = '';
-
-    public string $prefix = '';
-
-    public array $deps = [];
-
-    public int $licenseType = self::LICENSE_SPEC;
-
-    public const LICENSE_SPEC = 0;
-    public const LICENSE_APACHE2 = 1;
-    public const LICENSE_BSD = 2;
-    public const LICENSE_GPL = 3;
-    public const LICENSE_LGPL = 4;
-    public const LICENSE_MIT = 5;
-    public const LICENSE_PHP = 6;
-
-    public function __construct(string $name)
-    {
-        $this->name = $name;
-    }
-
-    public function withLicense(string $license, int $licenseType = self::LICENSE_SPEC): static
-    {
-        $this->license = $license;
-        $this->licenseType = $licenseType;
-        return $this;
-    }
-
-    public function withHomePage(string $homePage): static
-    {
-        $this->homePage = $homePage;
-        return $this;
-    }
-
-    public function withManual(string $manual): static
-    {
-        $this->manual = $manual;
-        return $this;
-    }
-
-    public function depends(string ...$libs): static
-    {
-        $this->deps += $libs;
-        return $this;
-    }
-
-    public function withMd5sum(string $md5sum): static
-    {
-        $this->md5sum = $md5sum;
-        return $this;
-    }
-}
-
-class Library extends Project
-{
-    public array $mirrorUrls = [];
-    public string $configure = '';
-    public string $ldflags = '';
-
-    public string $buildScript = '';
-    public string $makeOptions = '';
-    public string $makeVariables = '';
-    public string $makeInstallCommand = 'install';
-    public string $makeInstallOptions = '';
-    public string $beforeInstallScript = '';
-    public string $afterInstallScript = '';
-    public string $pkgConfig = '';
-    public array $pkgNames = [];
-    public string $prefix = '/usr';
-
-    public string $binPath = '';
-
-    public function withUrl(string $url): static
-    {
-        $this->url = $url;
-        return $this;
-    }
-    public function withMirrorUrl(string $url): static
-    {
-        $this->mirrorUrls[] = $url;
-        return $this;
-    }
-
-    public function withPrefix(string $prefix): static
-    {
-        $this->prefix = $prefix;
-        $this->withLdflags('-L' . $prefix . '/lib');
-        $this->withPkgConfig($prefix . '/lib/pkgconfig');
-        return $this;
-    }
-
-    public function getPrefix(): string
-    {
-        return $this->prefix;
-    }
-
-    public function withFile(string $file): static
-    {
-        $this->file = $file;
-        return $this;
-    }
-
-    public function withBuildScript(string $script): static
-    {
-        $this->buildScript = $script;
-        return $this;
-    }
-
-    public function withConfigure(string $configure): static
-    {
-        $this->configure = $configure;
-        return $this;
-    }
-
-    public function withLdflags(string $ldflags): static
-    {
-        $this->ldflags = $ldflags;
-        return $this;
-    }
-
-    public function withMakeVariables(string $variables): static
-    {
-        $this->makeVariables = $variables;
-        return $this;
-    }
-
-    public function withMakeOptions(string $makeOptions): static
-    {
-        $this->makeOptions = $makeOptions;
-        return $this;
-    }
-
-    public function withScriptBeforeInstall(string $script): static
-    {
-        $this->beforeInstallScript = $script;
-        return $this;
-    }
-
-    public function withScriptAfterInstall(string $script): static
-    {
-        $this->afterInstallScript = $script;
-        return $this;
-    }
-
-    public function withMakeInstallCommand(string $makeInstallCommand): static
-    {
-        $this->makeInstallCommand = $makeInstallCommand;
-        return $this;
-    }
-
-    public function withMakeInstallOptions(string $makeInstallOptions): static
-    {
-        $this->makeInstallOptions = $makeInstallOptions;
-        return $this;
-    }
-
-    public function withPkgConfig(string $pkgConfig): static
-    {
-        $this->pkgConfig = $pkgConfig;
-        return $this;
-    }
-
-    public function withPkgName(string $pkgName): static
-    {
-        $this->pkgNames[] = $pkgName;
-        return $this;
-    }
-
-    public function withBinPath(string $path): static
-    {
-        $this->binPath = $path;
-        return $this;
-    }
-}
-
-class Extension extends Project
-{
-    public string $options = '';
-    public string $peclVersion = '';
-
-    public function withOptions(string $options): static
-    {
-        $this->options = $options;
-        return $this;
-    }
-
-    public function withUrl(string $url): static
-    {
-        $this->url = $url;
-        return $this;
-    }
-
-    public function withPeclVersion(string $peclVersion): static
-    {
-        $this->peclVersion = $peclVersion;
-        return $this;
-    }
-}
-
 class Preprocessor
 {
     public const VERSION = '1.6';
@@ -511,7 +301,7 @@ class Preprocessor
         }
         // 下载失败
         if (!is_file($file) or filesize($file) == 0) {
-            throw new RuntimeException("Downloading file[$file] from url[$url] failed");
+            throw new RuntimeException("Downloading file[" . basename($file) . "] from url[$url] failed");
         }
         // 下载文件的 MD5 不一致
         if (!empty($md5sum) and !$this->checkFileMd5sum($file, $md5sum)) {
@@ -784,7 +574,7 @@ class Preprocessor
     public function execute(): void
     {
         if (empty($this->rootDir)) {
-            $this->rootDir = dirname(__DIR__);
+            $this->rootDir = dirname(__DIR__, 2);
         }
         if (empty($this->libraryDir)) {
             $this->libraryDir = $this->rootDir . '/pool/lib';
